requirements.txt
# GPTBots API 调用所需的基础依赖
requests>=2.28.0

# Streamlit 平台依赖
streamlit>=1.28.0

# 邮件处理依赖
email-validator>=1.3.0

# 文件处理依赖
python-magic>=0.4.27
markdown>=3.4.0
pandas>=1.5.0

# 其他工具依赖
pathlib>=1.0.1
streamlit-option-menu>=0.3.6
python-dotenv>=0.21.0